requirements.txt
# Core LLM + LangChain
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.8
openai>=1.12.0

# Database connectivity
mysql-connector-python==8.4.0
SQLAlchemy>=2.0.29

# Environment variable management
python-dotenv==1.0.1

# CLI / utilities (safe defaults)
pydantic>=2.6.0
tenacity>=8.2.3
